+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/controllers/data/SpatController.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.api.controllers;
+package us.dot.its.jpo.ode.api.controllers.data;
 
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
