+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/kafka/MapSocketForwardListener.java
@@ -7,27 +7,18 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.LineString;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.map.ProcessedMap;
-import us.dot.its.jpo.ode.api.controllers.StompController;
+import us.dot.its.jpo.ode.api.controllers.live.StompController;
 
 @Component
 @Slf4j
-@ConditionalOnProperty(
-    name = "enable.api",
-    havingValue = "true",
-    matchIfMissing = false
-)
+@ConditionalOnProperty(name = "enable.api", havingValue = "true", matchIfMissing = false)
 public class MapSocketForwardListener extends BaseSeekToEndListener {
 
     public MapSocketForwardListener(StompController stompController) {
         super(stompController);
     }
 
-    @KafkaListener(id = ListenerIds.MAP,
-            idIsGroup = false,
-            topics = "topic.ProcessedMap",
-            concurrency = "1",
-            containerFactory = "mapListenerContainerFactory",
-            autoStartup = "${conflict.monitor.api.kafka-consumers-always-on}")
+    @KafkaListener(id = ListenerIds.MAP, idIsGroup = false, topics = "topic.ProcessedMap", concurrency = "1", containerFactory = "mapListenerContainerFactory", autoStartup = "${conflict.monitor.api.kafka-consumers-always-on}")
     public void listen(ConsumerRecord<String, ProcessedMap<LineString>> record) {
         stompController.broadcastMap(record.value());
         log.trace("Received map with offset {}", record.offset());
