+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/kafka/SpatSocketForwardListener.java
@@ -6,28 +6,18 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 import us.dot.its.jpo.geojsonconverter.pojos.spat.ProcessedSpat;
-import us.dot.its.jpo.ode.api.controllers.StompController;
-
+import us.dot.its.jpo.ode.api.controllers.live.StompController;
 
 @Component
 @Slf4j
-@ConditionalOnProperty(
-    name = "enable.api",
-    havingValue = "true",
-    matchIfMissing = false
-)
+@ConditionalOnProperty(name = "enable.api", havingValue = "true", matchIfMissing = false)
 public class SpatSocketForwardListener extends BaseSeekToEndListener {
 
     public SpatSocketForwardListener(StompController stompController) {
         super(stompController);
     }
 
-    @KafkaListener(id = ListenerIds.SPAT,
-            idIsGroup = false,
-            topics = "topic.ProcessedSpat",
-            concurrency = "1",
-            containerFactory = "spatListenerContainerFactory",
-            autoStartup = "${conflict.monitor.api.kafka-consumers-always-on}")
+    @KafkaListener(id = ListenerIds.SPAT, idIsGroup = false, topics = "topic.ProcessedSpat", concurrency = "1", containerFactory = "spatListenerContainerFactory", autoStartup = "${conflict.monitor.api.kafka-consumers-always-on}")
     public void listen(ConsumerRecord<String, ProcessedSpat> record) {
         stompController.broadcastSpat(record.value());
         log.trace("Received spat with offset {}", record.offset());
