+++ services/intersection-api/api/src/test/java/us/dot/its/jpo/ode/api/MapTest.java
@@ -23,7 +23,7 @@ import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.LineString;
 import us.dot.its.jpo.geojsonconverter.pojos.geojson.map.ProcessedMap;
 import us.dot.its.jpo.ode.api.accessors.map.ProcessedMapRepository;
-import us.dot.its.jpo.ode.api.controllers.MapController;
+import us.dot.its.jpo.ode.api.controllers.data.MapController;
 import us.dot.its.jpo.ode.api.services.PermissionService;
 import us.dot.its.jpo.ode.mockdata.MockMapGenerator;
