+++ services/intersection-api/api/src/test/java/us/dot/its/jpo/ode/api/NotificationTest.java
@@ -36,7 +36,7 @@ import us.dot.its.jpo.ode.api.accessors.notifications.MapBroadcastRateNotificati
 import us.dot.its.jpo.ode.api.accessors.notifications.SignalGroupAlignmentNotificationRepo.SignalGroupAlignmentNotificationRepository;
 import us.dot.its.jpo.ode.api.accessors.notifications.SignalStateConflictNotification.SignalStateConflictNotificationRepository;
 import us.dot.its.jpo.ode.api.accessors.notifications.SpatBroadcastRateNotification.SpatBroadcastRateNotificationRepository;
-import us.dot.its.jpo.ode.api.controllers.NotificationController;
+import us.dot.its.jpo.ode.api.controllers.data.NotificationController;
 import us.dot.its.jpo.ode.api.services.PermissionService;
 import us.dot.its.jpo.ode.mockdata.MockNotificationGenerator;
 
@@ -46,257 +46,261 @@ import us.dot.its.jpo.ode.mockdata.MockNotificationGenerator;
 @AutoConfigureEmbeddedDatabase
 public class NotificationTest {
 
-    private final NotificationController controller;
+        private final NotificationController controller;
 
-    @MockBean
-    IntersectionReferenceAlignmentNotificationRepository intersectionReferenceAlignmentNotificationRepo;
+        @MockBean
+        IntersectionReferenceAlignmentNotificationRepository intersectionReferenceAlignmentNotificationRepo;
 
-    @MockBean
-    LaneDirectionOfTravelNotificationRepository laneDirectionOfTravelNotificationRepo;
+        @MockBean
+        LaneDirectionOfTravelNotificationRepository laneDirectionOfTravelNotificationRepo;
 
-    @MockBean
-    MapBroadcastRateNotificationRepository mapBroadcastRateNotificationRepo;
+        @MockBean
+        MapBroadcastRateNotificationRepository mapBroadcastRateNotificationRepo;
 
-    @MockBean
-    SignalGroupAlignmentNotificationRepository signalGroupAlignmentNotificationRepo;
+        @MockBean
+        SignalGroupAlignmentNotificationRepository signalGroupAlignmentNotificationRepo;
 
-    @MockBean
-    SignalStateConflictNotificationRepository signalStateConflictNotificationRepo;
+        @MockBean
+        SignalStateConflictNotificationRepository signalStateConflictNotificationRepo;
 
-    @MockBean
-    SpatBroadcastRateNotificationRepository spatBroadcastRateNotificationRepo;
+        @MockBean
+        SpatBroadcastRateNotificationRepository spatBroadcastRateNotificationRepo;
 
-    @MockBean
-    ConnectionOfTravelNotificationRepository connectionOfTravelNotificationRepo;
+        @MockBean
+        ConnectionOfTravelNotificationRepository connectionOfTravelNotificationRepo;
 
-    @MockBean
-    ActiveNotificationRepository activeNotificationRepo;
+        @MockBean
+        ActiveNotificationRepository activeNotificationRepo;
 
-    @MockBean
-    PermissionService permissionService;
+        @MockBean
+        PermissionService permissionService;
 
-    @Autowired
-    public NotificationTest(NotificationController controller) {
-        this.controller = controller;
-    }
+        @Autowired
+        public NotificationTest(NotificationController controller) {
+                this.controller = controller;
+        }
 
-    @Test
-    public void testConnectionOfTravelNotification() {
-        ConnectionOfTravelNotification notification = MockNotificationGenerator.getConnectionOfTravelNotification();
+        @Test
+        public void testConnectionOfTravelNotification() {
+                ConnectionOfTravelNotification notification = MockNotificationGenerator
+                                .getConnectionOfTravelNotification();
 
-        List<ConnectionOfTravelNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<ConnectionOfTravelNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(connectionOfTravelNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(connectionOfTravelNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<ConnectionOfTravelNotification>> result = controller
-                .findConnectionOfTravelNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<ConnectionOfTravelNotification>> result = controller
+                                .findConnectionOfTravelNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testIntersectionReferenceAlignmentNotification() {
-        IntersectionReferenceAlignmentNotification notification = MockNotificationGenerator
-                .getIntersectionReferenceAlignmentNotification();
+        @Test
+        public void testIntersectionReferenceAlignmentNotification() {
+                IntersectionReferenceAlignmentNotification notification = MockNotificationGenerator
+                                .getIntersectionReferenceAlignmentNotification();
 
-        List<IntersectionReferenceAlignmentNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<IntersectionReferenceAlignmentNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(intersectionReferenceAlignmentNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(intersectionReferenceAlignmentNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<IntersectionReferenceAlignmentNotification>> result = controller
-                .findIntersectionReferenceAlignmentNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<IntersectionReferenceAlignmentNotification>> result = controller
+                                .findIntersectionReferenceAlignmentNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testLaneDirectionOfTravelNotification() {
-        LaneDirectionOfTravelNotification notification = MockNotificationGenerator
-                .getLaneDirectionOfTravelNotification();
+        @Test
+        public void testLaneDirectionOfTravelNotification() {
+                LaneDirectionOfTravelNotification notification = MockNotificationGenerator
+                                .getLaneDirectionOfTravelNotification();
 
-        List<LaneDirectionOfTravelNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<LaneDirectionOfTravelNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(laneDirectionOfTravelNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(laneDirectionOfTravelNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<LaneDirectionOfTravelNotification>> result = controller
-                .findLaneDirectionOfTravelNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<LaneDirectionOfTravelNotification>> result = controller
+                                .findLaneDirectionOfTravelNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testMapBroadcastRateNotification() {
-        MapBroadcastRateNotification notification = MockNotificationGenerator.getMapBroadcastRateNotification();
+        @Test
+        public void testMapBroadcastRateNotification() {
+                MapBroadcastRateNotification notification = MockNotificationGenerator.getMapBroadcastRateNotification();
 
-        List<MapBroadcastRateNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<MapBroadcastRateNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(mapBroadcastRateNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(mapBroadcastRateNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<MapBroadcastRateNotification>> result = controller
-                .findMapBroadcastRateNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<MapBroadcastRateNotification>> result = controller
+                                .findMapBroadcastRateNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testSignalGroupAlignmentNotification() {
-        SignalGroupAlignmentNotification notification = MockNotificationGenerator.getSignalGroupAlignmentNotification();
+        @Test
+        public void testSignalGroupAlignmentNotification() {
+                SignalGroupAlignmentNotification notification = MockNotificationGenerator
+                                .getSignalGroupAlignmentNotification();
 
-        List<SignalGroupAlignmentNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<SignalGroupAlignmentNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(signalGroupAlignmentNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(signalGroupAlignmentNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<SignalGroupAlignmentNotification>> result = controller
-                .findSignalGroupAlignmentNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<SignalGroupAlignmentNotification>> result = controller
+                                .findSignalGroupAlignmentNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testSignalStateConflictNotification() {
-        SignalStateConflictNotification notification = MockNotificationGenerator.getSignalStateConflictNotification();
+        @Test
+        public void testSignalStateConflictNotification() {
+                SignalStateConflictNotification notification = MockNotificationGenerator
+                                .getSignalStateConflictNotification();
 
-        List<SignalStateConflictNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<SignalStateConflictNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(signalStateConflictNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(signalStateConflictNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<SignalStateConflictNotification>> result = controller
-                .findSignalStateConflictNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<SignalStateConflictNotification>> result = controller
+                                .findSignalStateConflictNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testSpatBroadcastRateNotification() {
-        SpatBroadcastRateNotification notification = MockNotificationGenerator.getSpatBroadcastRateNotification();
+        @Test
+        public void testSpatBroadcastRateNotification() {
+                SpatBroadcastRateNotification notification = MockNotificationGenerator
+                                .getSpatBroadcastRateNotification();
 
-        List<SpatBroadcastRateNotification> notifications = new ArrayList<>();
-        notifications.add(notification);
+                List<SpatBroadcastRateNotification> notifications = new ArrayList<>();
+                notifications.add(notification);
 
-        when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(notification.getIntersectionID(), "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(spatBroadcastRateNotificationRepo.find(notification.getIntersectionID(),
-                notification.getNotificationGeneratedAt() - 1,
-                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(spatBroadcastRateNotificationRepo.find(notification.getIntersectionID(),
+                                notification.getNotificationGeneratedAt() - 1,
+                                notification.getNotificationGeneratedAt() + 1, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<SpatBroadcastRateNotification>> result = controller
-                .findSpatBroadcastRateNotifications(
-                        notification.getIntersectionID(),
-                        notification.getNotificationGeneratedAt() - 1,
-                        notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<SpatBroadcastRateNotification>> result = controller
+                                .findSpatBroadcastRateNotifications(
+                                                notification.getIntersectionID(),
+                                                notification.getNotificationGeneratedAt() - 1,
+                                                notification.getNotificationGeneratedAt() + 1, false, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
-    @Test
-    public void testActiveNotification() {
-        List<Integer> allowedInteresections = new ArrayList<>();
-        allowedInteresections.add(null);
+        @Test
+        public void testActiveNotification() {
+                List<Integer> allowedInteresections = new ArrayList<>();
+                allowedInteresections.add(null);
 
-        when(permissionService.hasIntersection(null, "USER")).thenReturn(true);
-        when(permissionService.hasRole("USER")).thenReturn(true);
+                when(permissionService.hasIntersection(null, "USER")).thenReturn(true);
+                when(permissionService.hasRole("USER")).thenReturn(true);
 
-        SpatBroadcastRateNotification spatBroadcastRateNotification = MockNotificationGenerator
-                .getSpatBroadcastRateNotification();
-        SignalStateConflictNotification signalStateConflictNotification = MockNotificationGenerator
-                .getSignalStateConflictNotification();
-        SignalGroupAlignmentNotification signalGroupAlignmentNotification = MockNotificationGenerator
-                .getSignalGroupAlignmentNotification();
-        MapBroadcastRateNotification mapBroadcastRateNotification = MockNotificationGenerator
-                .getMapBroadcastRateNotification();
-        LaneDirectionOfTravelNotification laneDirectionOfTravelNotification = MockNotificationGenerator
-                .getLaneDirectionOfTravelNotification();
-        ConnectionOfTravelNotification connectionOfTravelNotification = MockNotificationGenerator
-                .getConnectionOfTravelNotification();
+                SpatBroadcastRateNotification spatBroadcastRateNotification = MockNotificationGenerator
+                                .getSpatBroadcastRateNotification();
+                SignalStateConflictNotification signalStateConflictNotification = MockNotificationGenerator
+                                .getSignalStateConflictNotification();
+                SignalGroupAlignmentNotification signalGroupAlignmentNotification = MockNotificationGenerator
+                                .getSignalGroupAlignmentNotification();
+                MapBroadcastRateNotification mapBroadcastRateNotification = MockNotificationGenerator
+                                .getMapBroadcastRateNotification();
+                LaneDirectionOfTravelNotification laneDirectionOfTravelNotification = MockNotificationGenerator
+                                .getLaneDirectionOfTravelNotification();
+                ConnectionOfTravelNotification connectionOfTravelNotification = MockNotificationGenerator
+                                .getConnectionOfTravelNotification();
 
-        List<Notification> notifications = new ArrayList<>();
-        notifications.add(spatBroadcastRateNotification);
-        notifications.add(signalStateConflictNotification);
-        notifications.add(signalGroupAlignmentNotification);
-        notifications.add(mapBroadcastRateNotification);
-        notifications.add(laneDirectionOfTravelNotification);
-        notifications.add(connectionOfTravelNotification);
+                List<Notification> notifications = new ArrayList<>();
+                notifications.add(spatBroadcastRateNotification);
+                notifications.add(signalStateConflictNotification);
+                notifications.add(signalGroupAlignmentNotification);
+                notifications.add(mapBroadcastRateNotification);
+                notifications.add(laneDirectionOfTravelNotification);
+                notifications.add(connectionOfTravelNotification);
 
-        PageRequest page = PageRequest.of(1, 1);
-        when(activeNotificationRepo.find(null,
-                null,
-                null, PageRequest.of(1, 1)))
-                .thenReturn(new PageImpl<>(notifications, page, 1L));
+                PageRequest page = PageRequest.of(1, 1);
+                when(activeNotificationRepo.find(null,
+                                null,
+                                null, PageRequest.of(1, 1)))
+                                .thenReturn(new PageImpl<>(notifications, page, 1L));
 
-        ResponseEntity<Page<Notification>> result = controller
-                .findActiveNotifications(
-                        null, null, null, 1, 1, false);
-        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(result.getBody().getContent()).isEqualTo(notifications);
-    }
+                ResponseEntity<Page<Notification>> result = controller
+                                .findActiveNotifications(
+                                                null, null, null, 1, 1, false);
+                assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
+                assertThat(result.getBody().getContent()).isEqualTo(notifications);
+        }
 
 }
